Fix integer related warnings in MSVC, minor refactoring (#2422)
- disable various MSVC warnings for a fully valid code;
- fix multiple warnings by casting data to a proper type and by making other minor code adjustments;
- filter out [useless](https://stackoverflow.com/questions/1453767/prevent-cl-exe-from-printing-the-compiled-source-file) MSVC information when building datatable;
- fix a possible bug in `check_slice_triple()` caused by the C++ operator precedence;
- use variable template for `RowIndex::NA`;
- switch Travis from `xenial` to `bionic` [environment](https://docs.travis-ci.com/user/languages/cpp/#gcc-on-linux) as `xenial` uses an old buggy gcc;

Closes #2413

+++ src/core/frame/repr/text_column.cc
@@ -262,7 +262,7 @@ tstring Data_TextColumn::_escape_string(const CString& str) const
           ch = ch0;
           break;
         }
-        remaining_width -= escaped.size();
+        remaining_width -= static_cast<int>(escaped.size());
         out << std::move(escaped);
       }
     }
